pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solark-cloud-ha"
version = "0.5.0"
description = "Home Assistant custom integration for Sol-Ark solar inverter systems via SolarkCloud API"
requires-python = ">=3.13"
license = "Apache-2.0"
dependencies = [
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["custom_components/solark_cloud"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=6",
    "pytest-asyncio>=0.24",
    "pytest-homeassistant-custom-component>=0.13",
    "respx>=0.22",
    "ruff>=0.11",
]

[tool.ruff]
target-version = "py312"
src = ["custom_components"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["custom_components/solark_cloud"]
branch = true

[tool.coverage.report]
fail_under = 90
show_missing = true
